# Optimizing performance for processing raw to events with GDC and TDC