--- a/sophiread/FastSophiread/include/tpx3_fast.h
+++ b/sophiread/FastSophiread/include/tpx3_fast.h
@@ -76,3 +76,7 @@ void update_tdc_timestamp(const char* char_array, const unsigned long long& gdc_
 
 void update_gdc_timestamp_and_timer_lsb32(const char* char_array, unsigned long& timer_lsb32,
                                           unsigned long long& gdc_timestamp);
+
+template <typename ForwardIter>
+void process_tpx3_packets(TPX3& tpx3h, ForwardIter bytes_begin, ForwardIter bytes_end, unsigned long& tdc_timestamp,
+                          unsigned long long int& gdc_timestamp, unsigned long& timer_lsb32, bool extract_hits = true);