# Description of Pull Request

This pull request optimize the overall performance of processing raw data (with TDC and GDC turned on) to neutron events.

Example results:
```bash
8cz in 🌐 cg1d-daq1 in mcpevent2hist/sophiread/build on  opt_raw2event_with_tgdc via △ v3.27.4 via 🅒 sophiread on ☁️   
❯ ./FastSophireadBenchmarks_raw2events.app ~/tmp/suann_socket_uvlight_5s_serval32.tpx3 
[2023-09-16 22:41:32.653] [info] File size: 4788526376 bytes
[2023-09-16 22:42:15.128] [info] Read raw data: 42.477315 s
[2023-09-16 22:42:15.535] [info] Single thread processing...
[2023-09-16 22:42:16.289] [info] Locate all headers: 0.754348 s
[2023-09-16 22:42:16.609] [info] Locate all gdc timestamps: 0.320256 s
[2023-09-16 22:43:02.250] [info] Get all hits: 45.640978999999994 s
[2023-09-16 22:43:03.720] [info] Total time: 46.715582999999995 s
[2023-09-16 22:43:03.720] [info] Number of hits: 596759877
[2023-09-16 22:43:03.720] [info] Single thread processing speed: 1.277432e+07 hits/s
[2023-09-16 22:43:16.070] [info] bad/total hits: 0/596759877
[2023-09-16 22:43:17.756] [info] Multi-thread processing...
[2023-09-16 22:43:25.640] [info] Number of hits: 596759877
[2023-09-16 22:43:25.640] [info] Multi-thread processing: 6.425933 s
[2023-09-16 22:43:25.640] [info] Multi-thread processing speed: 9.286743e+07 hits/s
[2023-09-16 22:43:37.983] [info] bad/total hits: 0/596759877

8cz in 🌐 cg1d-daq1 in mcpevent2hist/sophiread/build on  opt_raw2event_with_tgdc via △ v3.27.4 via 🅒 sophiread on ☁️   took 2m6s 
❯ ./FastSophireadBenchmarks_raw2events.app ~/tmp/suann_socket_uvlight_5s_serval32.tpx3 
[2023-09-16 22:43:44.941] [info] File size: 4788526376 bytes
[2023-09-16 22:43:45.656] [info] Read raw data: 0.7171339999999999 s
[2023-09-16 22:43:46.055] [info] Single thread processing...
[2023-09-16 22:43:46.816] [info] Locate all headers: 0.760506 s
[2023-09-16 22:43:47.134] [info] Locate all gdc timestamps: 0.318445 s
[2023-09-16 22:44:32.744] [info] Get all hits: 45.610209999999995 s
[2023-09-16 22:44:34.176] [info] Total time: 46.689161 s
[2023-09-16 22:44:34.176] [info] Number of hits: 596759877
[2023-09-16 22:44:34.176] [info] Single thread processing speed: 1.278155e+07 hits/s
[2023-09-16 22:44:46.593] [info] bad/total hits: 0/596759877
[2023-09-16 22:44:48.285] [info] Multi-thread processing...
[2023-09-16 22:44:56.275] [info] Number of hits: 596759877
[2023-09-16 22:44:56.275] [info] Multi-thread processing: 6.532737999999999 s
[2023-09-16 22:44:56.275] [info] Multi-thread processing speed: 9.134912e+07 hits/s
[2023-09-16 22:45:08.911] [info] bad/total hits: 0/596759877
```


## Purpose of work
<!--
Why has this work been done?
If there is no linked issue please provide appropriate context for this work.
-->

With added time stamp (TDC and GDC), additional time stamp sync is needed among TPX3 instances in order for multi-threading to generate correct results.
An additional low overhead step is added to scan the key/starting time stamps for each TPX3 so that subsequent processing can have correct TOF results.

<!-- If the original issue was raised by a user they should be named here.
NOTE: you can use @GITHUB_USERNAME to reference a user.
-->

## Summary of work
<!-- Please provide a short, high level description of the work that was done.
-->
1. Add new low overhead function to scan key/starting timestamp for each TPX3 header.
2. Update unit test and switch testing data to dataset with proper timestamp.
3. Update API calls in benchmark apps to use the new syntax.
4. Remove dead code.
5. Refactor the source code to reduce code duplications.
6. Update doc string for functions in `tpx3_fast.cpp` to indicate data flow direction.

## Additional detail of work
<!-- [Optional] If there is additional detail that is relevant to this PR, please provide it here.
-->

N/A

## Testing instructions

- Build the branch.
- Run all unit test with `ctest -V --output-on-failure`.
- Run benchmark app with dataset that has GDC and TDC turned.

<!-- Instructions for testing.
There should be sufficient instructions for someone unfamiliar with the application to test - unless a specific
reviewer is requested.
If instructions for replicating the fault are contained in the linked issue then it is OK to refer back to these.
-->

<!-- and fix #xxxx or close #xxxx xor resolves #xxxx 
NOTE: skip this part if not applicable to your PR.
-->